updated controller

+++ controllers/contactController.js
@@ -5,7 +5,7 @@ const Contact = require('../models/contactModel');
 //@route Get /api/contacts
 //@access private
 const getContacts = asyncHandler(async (req, res) => {
-  const contacts = await Contact.find();
+  const contacts = await Contact.find({ user_id: req.user.id });
   res.status(200).json(contacts);
 });
 
@@ -23,6 +23,7 @@ const createContact = asyncHandler(async (req, res) => {
     name,
     email,
     phone,
+    user_id: req.user.id,
   });
   res.status(201).json(contact);
 });
@@ -51,6 +52,11 @@ const updateContact = asyncHandler(async (req, res) => {
     throw new Error('Contact not found');
   }
 
+  if (contact.user_id.toString() !== req.user.id) {
+    req.status(403);
+    throw new Error('User dont have permission to update other contacts');
+  }
+
   const updatedContact = await Contact.findByIdAndUpdate(
     req.params.id,
     req.body,
@@ -70,6 +76,11 @@ const deleteContact = asyncHandler(async (req, res) => {
   if (!contact) {
     res.status(404);
     throw new Error('Contact not found');
+  }
+
+  if (contact.user_id.toString() !== req.user.id) {
+    req.status(403);
+    throw new Error('User dont have permission to update other contacts');
   }
 
   await Contact.deleteOne({ _id: req.params.id });
